noWalett observers

+++ src/main/java/org/erachain/gui/items/assets/TableModelItemAssetsFavorute.java
@@ -161,12 +161,14 @@ public class TableModelItemAssetsFavorute extends TableModelCls<Long, AssetCls> 
     }
 
     public void removeObservers() {
-        Controller.getInstance().wallet.database.getAssetFavoritesSet().deleteObserver(this);
+        if (Controller.getInstance().doesWalletDatabaseExists())
+            Controller.getInstance().wallet.database.getAssetFavoritesSet().deleteObserver(this);
     }
 
     public void addObservers() {
-
-        Controller.getInstance().wallet.database.getAssetFavoritesSet().addObserver(this);
+        if (Controller.getInstance().doesWalletDatabaseExists()) {
+            Controller.getInstance().wallet.database.getAssetFavoritesSet().addObserver(this);
+        }
     }
 
     @Override
